Move DrillDownTable files; restructure directory
