--- src/components/Table/helpers/SortIcon/sortHelpers.css
@@ -1,45 +1,0 @@
-/** Credit: cssicon.space */
-.previous-next-filled.icon {
-  margin-left: 2px;
-  margin-top: 3px;
-  width: 17px;
-  height: 14px;
-  transform: rotate(90deg);
-}
-
-.icon:before,
-.icon:after {
-  content: '';
-  position: absolute;
-  display: block;
-}
-
-.previous-next-filled.icon:before {
-  width: 0;
-  height: 0;
-  border-top: solid 7px transparent;
-  border-bottom: solid 7px transparent;
-  border-right: solid 7px #d2d2d2;
-}
-
-.previous-next-filled.icon::after {
-  left: 10px;
-  width: 0;
-  height: 0;
-  border-top: solid 7px transparent;
-  border-bottom: solid 7px transparent;
-  border-left: solid 7px #d2d2d2;
-}
-
-.previous-next-filled.icon.asc:before {
-  border-right: solid 7px #2bacff;
-}
-
-.previous-next-filled.icon.desc:after {
-  border-left: solid 7px #2bacff;
-}
-
-.icon-wrapper {
-  display: inline-block;
-  margin-left: 5px;
-}
