--- src/components/Table/helpers/pagination/pagination.css
@@ -1,11 +1,0 @@
-.page-sizes-text {
-  margin-right: 10px;
-}
-
-.pagination {
-  margin-top: 10px;
-  margin-bottom: 10px;
-  position: relative;
-  right: 0;
-  border: 1px solid red;
-}
