--- src/components/Table/table.css
@@ -1,31 +1,0 @@
-.drill-down-table {
-  width: 100%;
-  border-collapse: collapse;
-  background-color: #fff;
-  border-radius: 5px;
-  box-shadow: 0 2px 4px 0 #d7d7e0;
-  text-align: left;
-}
-
-.drill-down-table tbody {
-  line-height: 2em;
-}
-
-.drill-down-table td,
-.drill-down-table th {
-  border: 1px solid #ddd;
-  padding: 8px; /* can use to control the row hright too*/
-}
-
-.drill-down-table tr:nth-child(even) {
-  background-color: #f2f2f2;
-}
-
-.drill-down-table tr:hover {
-  background-color: #eee;
-}
-
-.drill-down-table th {
-  padding-top: 12px;
-  padding-bottom: 12px;
-}
